Fixing warnings caused by publicly exposing classes and methods.
