+++ src/Xunit.SkippableFact/Sdk/SkippableTestMessageBus.cs
@@ -4,10 +4,7 @@
 namespace Xunit.Sdk
 {
     using System;
-    using System.Collections.Generic;
     using System.Linq;
-    using System.Text;
-    using System.Threading.Tasks;
     using Validation;
     using Xunit.Abstractions;
 
@@ -21,6 +18,7 @@ namespace Xunit.Sdk
         /// The original message bus to which all messages should be forwarded.
         /// </summary>
         private readonly IMessageBus inner;
+        private bool isDisposed;
 
         /// <summary>
         /// Initializes a new instance of the <see cref="SkippableTestMessageBus"/> class.
@@ -37,21 +35,22 @@ namespace Xunit.Sdk
         }
 
         /// <summary>
-        /// Gets an array of full names to exception types that should be interpreted as a skip result.
-        /// </summary>
-        internal string[] SkippingExceptionNames { get; }
-
-        /// <summary>
         /// Gets the number of tests that have been dynamically skipped.
         /// </summary>
         public int SkippedCount { get; private set; }
+
+        /// <summary>
+        /// Gets an array of full names to exception types that should be interpreted as a skip result.
+        /// </summary>
+        internal string[] SkippingExceptionNames { get; }
 
         /// <summary>
         /// Disposes the inner message bus.
         /// </summary>
         public void Dispose()
         {
-            this.inner.Dispose();
+            this.Dispose(true);
+            GC.SuppressFinalize(this);
         }
 
         /// <inheritdoc />
@@ -87,6 +86,26 @@ namespace Xunit.Sdk
             }
 
             return this.inner.QueueMessage(message);
+        }
+
+        /// <summary>
+        /// The bulk of the clean-up code is implemented in Dispose(bool).
+        /// </summary>
+        /// <param name="disposing">If the managed resources should be disposed.</param>
+        protected virtual void Dispose(bool disposing)
+        {
+            if (this.isDisposed)
+            {
+                return;
+            }
+
+            if (disposing)
+            {
+                // free managed resources
+                this.inner.Dispose();
+            }
+
+            this.isDisposed = true;
         }
 
         private bool ShouldSkipException(string exceptionType) =>
